logthrsourcedrv: publish LogThreadedSourceWorker
The worker struct was kept private however various thread related actions
should be delegated to the concrete thread based implementations that
are inherently part of the worker class. We should implement a split
similar to what we already have on the destination side, e.g.
LogThreadedSourceWorker and LogThreadedSourceDriver.

This step publishes the private struct as a first step.
